+++ Frontend/src/utilities/hooks/useCoursesContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { CoursesContext } from '../context/course/coursesContext';
+import { CoursesContext } from '../context/course/CoursesContext';
 
 export function useCoursesContext() {
   return useContext(CoursesContext);
